chore: extract transformToPortfolioToken util
